refactor(runtime): move prompt_builder inline tests into their own file
The 1,014-LOC `#[cfg(test)] mod tests { ... }` block at the bottom of
`prompt_builder.rs` moves into `prompt_builder/tests.rs`. The parent
file declares it with `#[cfg(test)] mod tests;` so the suite stays
path-equivalent.

Dedent + de-nest pulls the previously-nested `#[test]` fns up one
level, keeping test names flat and `super::*` resolving to
prompt_builder (not prompt_builder::tests::tests).

prompt_builder.rs: 2,259 -> 1,243 LOC (-1,016). Now under the 2,000 LOC
file-size cap from #3710's acceptance criteria.

cargo check --tests + cargo clippy --lib both clean.

Refs #3710
